refactor: Move `util` to `internal` to avoid shadowing other libs
