+++ main/src/io/github/iltotore/iron/macros/union.scala
@@ -1,8 +1,8 @@
 package io.github.iltotore.iron.macros
 
 import io.github.iltotore.iron.{Constraint, Implication, ==>}
-import io.github.iltotore.iron.util.*
-import io.github.iltotore.iron.util.Validation.{Valid, Invalid}
+import io.github.iltotore.iron.internal.*
+import io.github.iltotore.iron.internal.Validation.{Valid, Invalid}
 import scala.quoted.*
 
 /**
